backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
dj-database-url==2.1.0
groq==0.4.2
httpx==0.27.2
python-dotenv==1.0.0
gunicorn==21.2.0
